requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
Pillow==7.2.0
pyfiglet==0.8.post1
pyperclip==1.8.1
pytube3==9.6.4
requests==2.24.0
typing-extensions==3.7.4.3
urllib3==1.25.10
